mstickerlib::database

Trait Database

pub trait Database {
    // Required methods
    async fn get(&self, hash: &Hash) -> Result<Option<String>>;
    async fn add(&self, hash: Hash, url: String) -> Result<()>;
}

Database which stores mappings from hashes to matrix media urls, to avoid duplicate uploads of the same file.

Required Methods

async fn get(&self, hash: &Hash) -> Result<Option<String>>

async fn add(&self, hash: Hash, url: String) -> Result<()>

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.