mstickerlib::matrix

Module sticker_formats

Current different StickerPack formats exists for Matrix. This mod store all current supported formats.

Modules

maunium

Stickerpacks for the maunium stickerpicker, which can be used at matrix clients whitch use the current sticker format, like Element and SchildiChat. The maunium stickerpicker does fully replace the default stickerpicker.

ponies

Ponies(MSC2545) are the new upcomming matrix sticker standard. They allow room and personal stickerpacks. This is already supported by many matrix clients like Neko, Cinny, Fluffychat and more. Keep in mind that ponies specification is not stable yet.