Struct mstickerlib::tg::ImportConfig

#[non_exhaustive]
pub struct ImportConfig<'a, D = DummyDatabase>where
    D: Database,{
    pub animation_format: Option<AnimationFormat>,
    pub database: Option<&'a D>,
    pub dry_run: bool,
}

additonal, optional configuration for importing stickers

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

animation_format: Option<AnimationFormat>

animaton format, to which animated sticker will be converted. If None original format will be used, this is propably not supported by matrix cilents.

database: Option<&'a D>

database to track, which files was already uploaded, to aviod duplicaded uploads of the same file

dry_run: bool

Do not upload anythink to matirx. WARNING: the generate stickerpack will not have valid matrix urls. Use this function only for testing and to prevent your homesever from being spammed with files while testing.